지도 api 17 - 랭킹등록, 랭킹조회 기능 추가

+++ CapstoneMap/app/src/main/java/com/example/capstonemap/locationUpdate/UserRecordDto.java
@@ -18,7 +18,7 @@ public class UserRecordDto {
     private Long id;
     private Long userId;
     private Long routeId;
-    private double elapsedTime;
+    private long elapsedTime;
     private List<Double[]> locationList;
 
     public UserRecordDto(Long userId, Long routeId){
