+++ springboot/src/main/java/com/example/capstoneMap/locationUpdate/UserRecordDto.java
@@ -19,17 +19,17 @@ public class UserRecordDto {
 	private Long id;
 	private Long userId;
 	private Long routeId;
-	private double elapsedTime;
+	private Long elapsedTime;
 	private List<Double[]> locationList;
 	
-	public UserRecordDto(Long userId, Long routeId, double elapsedTime, List<Double[]> locationList) {
+	public UserRecordDto(Long userId, Long routeId, Long elapsedTime, List<Double[]> locationList) {
 		this.userId=userId;
 		this.routeId=routeId;
 		this.elapsedTime=elapsedTime;
 		this.locationList=locationList;
 	}
 	
-	public UserRecordDto( double elapsedTime, List<Double[]> locationList) {
+	public UserRecordDto( long elapsedTime, List<Double[]> locationList) {
 		this.elapsedTime=elapsedTime;
 		this.locationList=locationList;
 	}
